requirements.txt
appdirs==1.4.4
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
coverage==5.5
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.4
django-adminlte3==0.1.6
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-embed-video==1.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
djangorestframework==3.12.2
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
idna==2.10
Inject==4.3.1
joblib==1.0.1
list-print==1.0.1
lxml==4.6.2
mercury==0.1
MouseInfo==0.1.3
mysqlclient==2.0.3
numpy==1.20.3
opencv-python==4.5.2.54
opencv-python-headless==4.5.2.54
Pillow==8.2.0
psycopg2==2.9.1
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
PySDL2==0.9.7
pysdl2-dll==2.0.14.post1
PyTweening==1.0.3
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
serializers==0.2.4
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.61.1
urllib3==1.26.5
virtualenv==20.2.2
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
